@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-top-loading-bar": "^2.0.1",
-    "web-vitals": "^1.1.2"
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
+    "react-top-loading-bar": "^2.3.1",
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
