@@ -19,7 +19,7 @@ const NewsItem = (props)=> {
                         <h5 className="card-title">{title}  </h5>
                         <p className="card-text">{description}</p>
                         <p className="card-text"><small className="text-muted">By {!author ? "Unknown" : author} on  {new Date(date).toGMTString()}</small></p>
-                        <a rel="noreferrer" href={newsUrl} target="_blank" className="btn btn-sm btn-dark">Read More</a>
+                        <a href={newsUrl} target="_blank" rel="noopener noreferrer" className="btn btn-sm btn-dark">Read More</a>
                     </div>
                 </div>
             </div>
